% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_export_pdf.R
\name{auto_export_ggplot}
\alias{auto_export_ggplot}
\title{Automatic save ggplot as PDF}
\usage{
auto_export_ggplot(
  dir,
  file,
  include_unnamed_chunk = FALSE,
  env = parent.frame()
)
}
\arguments{
\item{dir}{The directory to save the file}

\item{file}{The single file used to save all pdf}

\item{env}{The environment to register this method. Usually is the caller env.}
}
\description{
Automatic save ggplot as PDF
}
\details{
This function is called for its side-effect.

It defines a custom \code{\link[knitr:knit_print]{knitr::knit_print()}} method for object of class \code{gg},
i.e. \code{ggplot} objects. This custom print function will save the all \code{ggplot}
That is printed to the designated directory.

By default, only chunk with a name will be saved, and the  pdf file is
named after the chunk name. One can also use \code{include_unnamed_chunk = TRUE}.
In this case, they will be named something like \verb{unnamed-chunk-2.pdf}.
}
