% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pdf.R
\name{export_pdf}
\alias{export_pdf}
\alias{export_cairo_pdf}
\title{Export a list of figures to a single PDF}
\usage{
export_pdf(x, file, width, height, ...)

export_cairo_pdf(x, file, width, height, ...)
}
\arguments{
\item{x}{A list of object that can be understand by the pdf device.}

\item{file}{The output file name.}

\item{width}{The width of each page in the final the PDF.}

\item{height}{The height of each page in the final the PDF.}

\item{...}{Additional argument passed to the \code{\link[=pdf]{pdf()}} or \code{\link[=cairo_pdf]{cairo_pdf()}} device.}
}
\value{
Silently return the list of input object.
}
\description{
Export a list of figures to a single PDF
}
\details{
This function can be used to save multiple objects (usually plots) into
a single PDF file, with different dimensions.

A scalar can be used for the width and height argument.
In this case, this width/height will be applied to all pages
of the final PDF file.
}
